pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devex-metrics"
version = "0.1.0"
description = "GitHub Organization DevOps Metrics Analyzer - Calculate maturity scores based on State of DevOps 2026 benchmarks"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DevEx Metrics Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["devops", "metrics", "github", "dora", "maturity", "analysis"]

dependencies = [
    "typer[all]>=0.9.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "plotly>=5.18.0",
    "pydantic>=2.5.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
devex-metrics = "devex_metrics.cli:app"

[project.urls]
Homepage = "https://github.com/devex-metrics/devex-metrics"
Documentation = "https://github.com/devex-metrics/devex-metrics#readme"
Repository = "https://github.com/devex-metrics/devex-metrics"
"Bug Tracker" = "https://github.com/devex-metrics/devex-metrics/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["devex_metrics*"]

[tool.setuptools.package-data]
devex_metrics = ["reporters/templates/**/*.html", "reporters/templates/**/*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests for CLI",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "plotly.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true
